feat: 검색 결과가 없는 경우 UI 변경
